Merge `scigen-manual` and `scigen-manual-downsampled`. Some initial consolidation with `scigen-manual-old`.

+++ src/main/java/authoringassistant/Program.java
@@ -26,6 +26,7 @@ import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
+import java.util.stream.Stream;
 
 import static authoringassistant.variable.Variables.Flat.expandVariables;
 
@@ -386,7 +387,6 @@ public class Program {
         final String websitesRoot = "website/authoring-assistant/";
         final Path fileName = testCasePath.getFileName();
         Path websiteName = testCasePath.getParent().getFileName();
-        logger.info(STR."Adding page to website \{websiteName}");
         String websiteRoot = STR."\{websitesRoot}\{websiteName}/";
         String page = STR."\{websiteRoot}\{fileName}";
         Files.createDirectories(Path.of(page));
@@ -419,22 +419,22 @@ public class Program {
         writeFluidFiles(STR."\{websiteRoot}fluid/", STR."\{fileName}.fld", paragraph.toFluidSyntax(false), datasets, _loadedDatasets, imports, _loadedImports, code);
     }
 
-    public static void cleanWebsiteFolders(String path) {
+    public static void cleanWebsiteFolders(String path) throws IOException {
         Path directoryPath = Paths.get(path);
 
-        try (var paths = Files.walk(directoryPath)) {
+        if (Files.notExists(directoryPath)) {
+            Files.createDirectories(directoryPath);
+        } else {
+            final Stream<Path> paths = Files.walk(directoryPath);
             paths.filter(p -> !p.equals(directoryPath) && Files.isDirectory(p) && !Files.isSymbolicLink(p))
-                    .forEach(dir -> {
-                        try {
-                            FileUtils.deleteDirectory(dir.toFile());
-                        } catch (IOException e) {
-                            logger.info("Error during clean of website folder");
-                            throw new RuntimeException(e);
-                        }
-                    });
-        } catch (IOException e) {
-            logger.info("Error during clean of website folder");
-            throw new RuntimeException(e);
+                 .forEach(dir -> {
+                     try {
+                         FileUtils.deleteDirectory(dir.toFile());
+                     } catch (IOException e) {
+                         logger.info("Error during clean of website folder");
+                         throw new RuntimeException(e);
+                     }
+                 });
         }
     }
